separating utils_device_copyable.h
