+++ test/general/implementation_details/device_copyable.pass.cpp
@@ -24,7 +24,7 @@ using namespace TestUtils;
 
 #if TEST_DPCPP_BACKEND_PRESENT
 
-#    include "support/utils_sycl.h"
+#    include "support/utils_device_copyable.h"
 #    include "sycl.hpp"
 
 void
