+++ test/parallel_api/iterator/transform_iterator.pass.cpp
@@ -27,7 +27,7 @@ using namespace TestUtils;
 
 #if TEST_DPCPP_BACKEND_PRESENT
 
-#include "support/utils_sycl.h"
+#include "support/utils_device_copyable.h"
 
 DEFINE_TEST(test_copy)
 {
